api/requirements.txt
flask
psycopg2-binary==2.9.11
python-dotenv